Make code SOLID
