+++ config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   # root "articles#index"
   post 'pay', to: 'mpesas#stkpush'
-  post 'payment_query', to: 'mpesas#stkquery'
-  post 'b2c', to: 'mpesas#b2c'
-  post '/callback', to: 'mpesas#mpesa_callback'
+  post 'payment_query', to: 'callbacks#stkquery'
+  post 'b2c', to: 'b2c_transactions#b2c'
+  post 'callback', to: 'callbacks#mpesa_callback'
 end
